requirement.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
bootstrap4==0.1.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.3
crispy-forms-gds==0.2.1
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.1
django-bootstrap3==14.2.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.1.1
django-heroku==0.3.1
django-import-export==2.4.0
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
isort==5.4.2
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pylint==2.6.0
pytz==2020.1
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
sqlparse==0.3.1
tablib==3.0.0
toml==0.10.1
urllib3==1.26.3
virtualenv==20.2.2
whitenoise==5.2.0
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
xxx==0.0.1
